+++ src/lib/hooks/useObligations.ts
@@ -15,8 +15,18 @@ export function useObligations() {
   const { obligations, loading, error } = useAppSelector(
     state => state.obligations,
   );
+  const { user } = useAppSelector(state => state.auth);
+
+  const getUserObligation = (obligationId: string) => {
+    return obligations.find(
+      obligation => obligation.obligationId === obligationId,
+    );
+  };
 
   const fetchObligations = async () => {
+    if (loading) {
+      throw new Error("Already fetching obligations");
+    }
     dispatch(setLoading(true));
     try {
       const response = await axios.get("/api/obligation");
@@ -24,25 +34,42 @@ export function useObligations() {
       dispatch(setError(null));
     } catch (err: any) {
       dispatch(setError(err.message || "Error fetching obligations"));
+      throw err;
     } finally {
       dispatch(setLoading(false));
     }
   };
 
+  const setObligations = (obligations: Obligation[]) => {
+    dispatch(setObligationsAction(obligations));
+  };
+
   const createObligation = async (obligationData: CreateObligation) => {
+    if (loading) {
+      throw new Error("Already creating obligation");
+    }
+
     dispatch(setLoading(true));
     try {
-      const response = await axios.post("/api/obligation", obligationData);
+      const response = await axios.post("/api/obligation", {
+        ...obligationData,
+        userId: user?.userId,
+      });
       dispatch(createObligationAction(response.data.result));
       dispatch(setError(null));
     } catch (err: any) {
       dispatch(setError(err.message || "Error creating obligation"));
+      throw err;
     } finally {
       dispatch(setLoading(false));
     }
   };
 
   const updateObligation = async (obligationData: Obligation) => {
+    if (loading) {
+      throw new Error("Already updating obligation");
+    }
+
     dispatch(setLoading(true));
     try {
       const response = await axios.patch(
@@ -53,12 +80,16 @@ export function useObligations() {
       dispatch(setError(null));
     } catch (err: any) {
       dispatch(setError(err.message || "Error updating obligation"));
+      throw err;
     } finally {
       dispatch(setLoading(false));
     }
   };
 
   const deleteObligation = async (obligationId: string) => {
+    if (loading) {
+      throw new Error("Already deleting obligation");
+    }
     dispatch(setLoading(true));
     try {
       await axios.delete(`/api/obligation/${obligationId}`);
@@ -66,6 +97,7 @@ export function useObligations() {
       dispatch(setError(null));
     } catch (err: any) {
       dispatch(setError(err.message || "Error deleting obligation"));
+      throw err;
     } finally {
       dispatch(setLoading(false));
     }
@@ -75,6 +107,8 @@ export function useObligations() {
     obligations,
     loading,
     error,
+    setObligations,
+    getUserObligation,
     fetchObligations,
     createObligation,
     updateObligation,
